+++ Dip/DipTests/RuntimeArgumentsTests.swift
@@ -1,9 +1,25 @@
 //
-//  RuntimeArgumentsTests.swift
-//  DipTests
+// Dip
 //
-//  Created by Ilya Puchka on 04.11.15.
-//  Copyright © 2015 AliSoftware. All rights reserved.
+// Copyright (c) 2015 Olivier Halligon <olivier@halligon.net>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 //
 
 import XCTest
@@ -34,13 +50,13 @@ class RuntimeArgumentsTests: XCTestCase {
   func testThatItResolvesInstanceWithOneArgument() {
     //given
     let arg1 = 1
-    container.register { (a1: Int) -> Service in
+    container.register(factory: { (a1: Int) -> Service in
       XCTAssertEqual(a1, arg1)
       return ServiceImp1()
-    }
+    })
     
     //when
-    let service = container.resolve(withArguments: arg1) as Service
+    let service = try! container.resolve(withArguments: arg1) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -56,7 +72,7 @@ class RuntimeArgumentsTests: XCTestCase {
     }
     
     //when
-    let service = container.resolve(withArguments: arg1, arg2) as Service
+    let service = try! container.resolve(withArguments: arg1, arg2) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -72,7 +88,7 @@ class RuntimeArgumentsTests: XCTestCase {
     }
     
     //when
-    let service = container.resolve(withArguments: arg1, arg2, arg3) as Service
+    let service = try! container.resolve(withArguments: arg1, arg2, arg3) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -89,7 +105,7 @@ class RuntimeArgumentsTests: XCTestCase {
     }
     
     //when
-    let service = container.resolve(withArguments: arg1, arg2, arg3, arg4) as Service
+    let service = try! container.resolve(withArguments: arg1, arg2, arg3, arg4) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -107,7 +123,7 @@ class RuntimeArgumentsTests: XCTestCase {
     }
     
     //when
-    let service = container.resolve(withArguments: arg1, arg2, arg3, arg4, arg5) as Service
+    let service = try! container.resolve(withArguments: arg1, arg2, arg3, arg4, arg5) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -126,7 +142,7 @@ class RuntimeArgumentsTests: XCTestCase {
     }
     
     //when
-    let service = container.resolve(withArguments: arg1, arg2, arg3, arg4, arg5, arg6) as Service
+    let service = try! container.resolve(withArguments: arg1, arg2, arg3, arg4, arg5, arg6) as Service
     
     //then
     XCTAssertTrue(service is ServiceImp1)
@@ -139,8 +155,8 @@ class RuntimeArgumentsTests: XCTestCase {
     container.register { (a1: Int, a2: Int) in ServiceImp2() as Service }
     
     //when
-    let service1 = container.resolve(withArguments: arg1) as Service
-    let service2 = container.resolve(withArguments: arg1, arg2) as Service
+    let service1 = try! container.resolve(withArguments: arg1) as Service
+    let service2 = try! container.resolve(withArguments: arg1, arg2) as Service
     
     //then
     XCTAssertTrue(service1 is ServiceImp1)
@@ -150,12 +166,12 @@ class RuntimeArgumentsTests: XCTestCase {
   func testThatItRegistersDifferentFactoriesForDifferentTypesOfArguments() {
     //given
     let arg1 = 1, arg2 = "string"
-    container.register { (a1: Int) in ServiceImp1() as Service }
-    container.register { (a1: String) in ServiceImp2() as Service }
+    container.register(factory: { (a1: Int) in ServiceImp1() as Service })
+    container.register(factory: { (a1: String) in ServiceImp2() as Service })
     
     //when
-    let service1 = container.resolve(withArguments: arg1) as Service
-    let service2 = container.resolve(withArguments: arg2) as Service
+    let service1 = try! container.resolve(withArguments: arg1) as Service
+    let service2 = try! container.resolve(withArguments: arg2) as Service
     
     //then
     XCTAssertTrue(service1 is ServiceImp1)
@@ -169,8 +185,8 @@ class RuntimeArgumentsTests: XCTestCase {
     container.register { (a1: String, a2: Int) in ServiceImp2() as Service }
     
     //when
-    let service1 = container.resolve(withArguments: arg1, arg2) as Service
-    let service2 = container.resolve(withArguments: arg2, arg1) as Service
+    let service1 = try! container.resolve(withArguments: arg1, arg2) as Service
+    let service2 = try! container.resolve(withArguments: arg2, arg1) as Service
     
     //then
     XCTAssertTrue(service1 is ServiceImp1)
@@ -181,11 +197,11 @@ class RuntimeArgumentsTests: XCTestCase {
     //given
     let arg1 = 1, arg2 = 2
     container.register { (a1: Int, a2: Int) in ServiceImp1() as Service }
-    let service1 = container.resolve(withArguments: arg1, arg2) as Service
+    let service1 = try! container.resolve(withArguments: arg1, arg2) as Service
     
     //when
     container.register { (a1: Int, a2: Int) in ServiceImp2() as Service }
-    let service2 = container.resolve(withArguments: arg1, arg2) as Service
+    let service2 = try! container.resolve(withArguments: arg1, arg2) as Service
     
     //then
     XCTAssertTrue(service1 is ServiceImp1)
@@ -201,11 +217,11 @@ class RuntimeArgumentsTests: XCTestCase {
     
     //when
     let url: NSURL = NSURL(string: "http://example.com")!
-    let service1 = container.resolve(withArguments: 80, url) as Service
-    let service2 = container.resolve(withArguments: 80, NSURL(string: "http://example.com")) as Service
+    let service1 = try! container.resolve(withArguments: 80, url) as Service
+    let service2 = try! container.resolve(withArguments: 80, NSURL(string: "http://example.com")) as Service
     
-    let service3 = container.resolve(withArguments: 80, NSURL(string: "http://example.com")! as NSURL!) as Service
-    let service4 = container.resolve(withArguments: 80, NSURL(string: "http://example.com")!) as Service
+    let service3 = try! container.resolve(withArguments: 80, NSURL(string: "http://example.com")! as NSURL!) as Service
+    let service4 = try! container.resolve(withArguments: 80, NSURL(string: "http://example.com")!) as Service
     
     //then
     XCTAssertEqual(service1.getServiceName(), name1)